Works correctly now.
User was able to add more than one decimal point ("1.2.3.4.5.").  This has been corrected.  

Also had to remove the `.startswith("0")` lines in `insert_to_entry_box` because it prevented the user from inputting equations like "0.5+1".  This does mean the user is able to enter things like "1+02" again, which will result in an error.